migrate doi creation js code to a stimulus controller
